Changed folder structure to allow for future growth.
