+++ scripts/entities/spritesheet.py
